requirements.txt
djangorestframework>=3.9.0,<3.10.0
Django==2.2.6
gunicorn==20.0.4
psycorg2-binary==2.8.3
flake8

